Detect late addition of CUDA source files.
If target_sources is called explicitly the target is turned into
an RDC executable only at the next call and to
cuda_rdc_target_link_libraries and there we need to make sure to
turn on the RDC linking.

Note: if the same happens for a library, it might not yet be
supported either.